pyproject.toml
[project]
name = "blockchain-llm-agent-server"
version = "0.1.0"
description = ""
authors = [{ name = "tjjd4" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "ollama (>=0.4.7,<0.5.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "web3 (>=7.9.0,<8.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "ag2[mcp,ollama,openai]>=0.8.5",
    "streamlit>=1.44.1",
]
